Fix bcpid optimization level and compile errors
Summary: Set default optimization level to O3 and debug optimization level to Og. Enable Wextra and Werror.

Test Plan: make

Reviewers: ali, tavianator

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D795

+++ bcpiquery/elfutil.cc
@@ -183,7 +183,7 @@ load_sections(Elf *e, size_t shnum)
 		i++;
 	} while ((scn = elf_nextscn(e, scn)) != NULL);
 
-	if (i != shnum) {
+	if ((unsigned int)i != shnum) {
 		printf("error i = %d, shnum = %zu\n", i, shnum);
 	}
